requirements.txt
#django == 4.0.5
pillow
django-admin-interface
django-crispy-forms